Fix a bug in showing insertion failure alert dialog
If the insertion operation fails, the user won't be informed
about the failure. They'll be alerted only if the insertion
is successful. That's because the control flow never gets to
the 'else' clause. Actually it always checks if the result
is successful. This patch fixes this problem by applying
a more general concept called 'result' as the key and then
the value which can be either 'success' or any message that
the database generates when it cannot perform the insertion
operation and as may be expected it's treated as failure.
In fact API's bad response structure caused the bug and it
was also changed on the server side.